+++ R/utils-expr.R
@@ -423,7 +423,7 @@ translate <- function(
 
       user_defined <- get_globenv_functions()
       known_ops <- c("+", "-", "*", "/", ">", ">=", "<", "<=", "==", "!=",
-                     "&", "|", "!")
+                     "&", "|", "!", "%%", "%/%")
       fn_names <- add_pkg_suffix(name, known_ops, user_defined)
       name <- fn_names$name_to_eval
       is_known <- is_function_known(name)