+++ tests/tinytest/test_funs_math.R
@@ -4,7 +4,7 @@ using("tidypolars")
 test_df <- data.frame(
   x1 = c("a", "a", "b", "a", "c"),
   x2 = c(2, 1, 5, 3, 1),
-  value = sample(1:5),
+  value = sample(11:15),
   value_trigo = seq(0, 0.4, 0.1),
   value_mix = -2:2
 )
@@ -86,4 +86,24 @@ expect_equal(
 expect_error(
   test |> summarize(foo = diff(value, differences = 2)),
   "doesn't support"
+)
+
+# %% and %/%
+
+expect_equal(
+  test |>
+    mutate(foo = value %% 10) |>
+    pull(foo),
+  test_df |>
+    mutate(foo = value %% 10) |>
+    pull(foo)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = value %/% 10) |>
+    pull(foo),
+  test_df |>
+    mutate(foo = value %/% 10) |>
+    pull(foo)
 )